V1.2

+++ app/controllers/admin.php
@@ -11,11 +11,15 @@ require_once(MODEL_PATH . '/table_niveau.php');
 require_once(MODEL_PATH . '/table_prefixetel.php');
 require_once(MODEL_PATH . '/table_seanceclient.php');
 require_once(MODEL_PATH . '/table_seancecoach.php');
+require_once(MODEL_PATH . '/table_admin.php');
+require_once(MODEL_PATH . '/table_typeabonnement.php');
 
 //----------------------------- show statistics from database (about clients and coachs) -----------------------------
 
 function index () // admin home page, get statistics from database
 {
+	checkAdminConnexion();
+
 	$local_data = ['presentation_message' => 'Statistiques DigitalSport', // message de presentation
 				   ['Clients :',countClient()], // ['message',countFunction]
 				   ['Coachs :',countCoach()],
@@ -34,6 +38,7 @@ function index () // admin home page, get statistics from database
 
 function show_all_coachs_profile () // show all coach profile
 {
+	checkAdminConnexion();
 	$result = showAllCoachsProfile();
 
 	if (empty($result)){ // check if requested coach exists
@@ -47,6 +52,7 @@ function show_all_coachs_profile () // show all coach profile
 
 function show_all_clients_profile () // show all client profile
 {
+	checkAdminConnexion();
 	$result = showAllClientsProfile();
 
 	if (empty($result)){ // check if requested client exists
@@ -63,6 +69,7 @@ function show_all_clients_profile () // show all client profile
 
 function show_contact_message () // show contact message (from client)
 {
+	checkAdminConnexion();
 	$local_data = showContactMessage();
 
 	display_view('admin/show_contact_message', $local_data); // array format needed
@@ -73,6 +80,7 @@ function show_contact_message () // show contact message (from client)
 
 function delete_contact_message () // delete contact message (from database)
 {
+	checkAdminConnexion();
 	deleteContactMessageProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=show_contact_message'); // redirect on the view to show contact message with success message
@@ -83,6 +91,7 @@ function delete_contact_message () // delete contact message (from database)
 
 function delete_day_contact_message () // delete contact message for a specific day (from database)
 {
+	checkAdminConnexion();
 	deleteDayContactMessageProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=show_contact_message'); // redirect on the view to show contact message with success message
@@ -94,6 +103,7 @@ function delete_day_contact_message () // delete contact message for a specific 
 
 function add_exercise () // show form to add exercises
 {
+	checkAdminConnexion();
 	$local_data = getMaterial(); // get all material to choose
 
 	display_view('admin/add_exercise', ['presentation_message' => 'Exercice à ajouter :', $local_data]);
@@ -101,6 +111,7 @@ function add_exercise () // show form to add exercises
 
 function add_exercise_proceed () // proceeds to the exercise addition
 {
+	checkAdminConnexion();
 	addExerciseProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=show_exercise');
@@ -112,6 +123,7 @@ function add_exercise_proceed () // proceeds to the exercise addition
 
 function show_exercise () // show exercise
 {
+	checkAdminConnexion();
 	$local_data = getExercise(); // get all exercise to show
 
 	display_view('admin/show_exercise', ['presentation_message' => 'Liste des exercices :', 'success_message' => '', $local_data]);
@@ -121,6 +133,7 @@ function show_exercise () // show exercise
 
 function update_exercise () // show exercise to delete
 {
+	checkAdminConnexion();
 	$local_data = [getExerciseInformation(),getOrderedMaterial()]; // get all exercise to update
 
 	display_view('admin/update_exercise', ['presentation_message' => 'Exercice à modifier :', $local_data]);
@@ -128,7 +141,7 @@ function update_exercise () // show exercise to delete
 
 function update_exercise_proceed () // proceeds to the exercise update
 {
-
+	checkAdminConnexion();
 	$oldInfo = getExerciseInformation(); // get old information from database
 	$newInfo = getExerciseModification(); // get new information from website (with update)
 	
@@ -152,6 +165,7 @@ function update_exercise_proceed () // proceeds to the exercise update
 
 function delete_exercise_proceed () // proceeds to the exercise deletion
 {
+	checkAdminConnexion();
 	deleteExerciseProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=show_exercise');
@@ -161,8 +175,10 @@ function delete_exercise_proceed () // proceeds to the exercise deletion
 
 //----------------------------- update material -----------------------------
 
-function update_material (){ // show form to update material
+function update_material () // show form to update material
+{ 
 
+	checkAdminConnexion();
 	$local_data = getMaterial();
 
 	display_view('admin/update_material', ['presentation_message' => 'Modification matériel :', 'success_message' => '', $local_data]);
@@ -173,6 +189,7 @@ function update_material (){ // show form to update material
 
 function add_material_proceed () // proceeds to the material addition
 {
+	checkAdminConnexion();
 	addMaterialProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=update_material');
@@ -184,6 +201,7 @@ function add_material_proceed () // proceeds to the material addition
 
 function delete_material_proceed () // proceeds to the material deletion
 {
+	checkAdminConnexion();
 	deleteMaterialProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=update_material');
@@ -195,19 +213,21 @@ function delete_material_proceed () // proceeds to the material deletion
 
 function update_client_profile () // show all the informations about a client to update them
 {
-	$local_data = [getClientProfile(),getOrderedPrefixPhone(),getOrderedLevel(), 'action_message' => 'client', 'presentation_message' => 'Modification profil client :' ];
+	checkAdminConnexion();
+	$local_data = [getClientProfile(),getOrderedPrefixPhone(),getOrderedLevel(), getOrderedSubsciption(), 'action_message' => 'client', 'presentation_message' => 'Modification profil client :' ];
 
 	display_view('admin/update_user_profile', $local_data);
 }
 
 function update_client_profile_proceed () // proceeds to the update of a client profile
 {
+	checkAdminConnexion();
 	$oldInfo = getClientProfile(); // get old information from database
 	$newInfo = getModificationClientProfile(); // get new information from website (with update)
 	
 
 	$array_diff = array_diff_assoc($newInfo,$oldInfo[0]); // get the information to update by comparing old and new information
-	$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(), 'action_message'=>'client','presentation_message'=>'Modification profil client :','success_message' => 'Le profil du client à été édité avec succès ! 💪'];
+	$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(), getOrderedSubsciption(), 'action_message'=>'client','presentation_message'=>'Modification profil client :','success_message' => 'Le profil du client à été édité avec succès ! 💪'];
 
 	if (!empty($array_diff)) // check whether the information has been changed
 	{
@@ -216,7 +236,7 @@ function update_client_profile_proceed () // proceeds to the update of a client 
 		}
 
 		if (!empty($existingEmail)){
-			$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(),'action_message'=>'client', 'presentation_message'=>'Modification profil client :', 'error_message' => 'Cet email est déja utilisé par un autre client.'];
+			$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(), getOrderedSubsciption(),'action_message'=>'client', 'presentation_message'=>'Modification profil client :', 'error_message' => 'Cet email est déja utilisé par un autre client.'];
 			display_view('admin/update_user_profile', $local_data);
 			exit;
 		}
@@ -226,7 +246,7 @@ function update_client_profile_proceed () // proceeds to the update of a client 
 		}
 
 		if (!empty($existingPseudo)){
-			$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(),'action_message'=>'client', 'presentation_message'=>'Modification profil client :', 'error_message' => 'Ce pseudo est déja utilisé par un autre client.'];
+			$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(), getOrderedSubsciption(),'action_message'=>'client', 'presentation_message'=>'Modification profil client :', 'error_message' => 'Ce pseudo est déja utilisé par un autre client.'];
 			display_view('admin/update_user_profile', $local_data);
 			exit;
 		}
@@ -235,7 +255,7 @@ function update_client_profile_proceed () // proceeds to the update of a client 
 
 		display_view('admin/update_user_profile', $local_data);
 	} else {
-		$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(),'action_message'=>'client', 'presentation_message'=>'Modification profil client :', 'error_message' => "Aucune information n'a été modifiée"];
+		$local_data = [[$newInfo],getOrderedPrefixPhone(),getOrderedLevel(), getOrderedSubsciption(),'action_message'=>'client', 'presentation_message'=>'Modification profil client :', 'error_message' => "Aucune information n'a été modifiée"];
 		display_view('admin/update_user_profile', $local_data);
 	}
 
@@ -245,6 +265,7 @@ function update_client_profile_proceed () // proceeds to the update of a client 
 
 function delete_client_profile_proceed () // proceeds the client profile deletion
 {
+	checkAdminConnexion();
 	deleteClientProfileProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=show_all_clients_profile'); // redirect on the view to show client profile with success message
@@ -256,6 +277,7 @@ function delete_client_profile_proceed () // proceeds the client profile deletio
 
 function update_coach_profile () // show all the informations about a coach to update them
 {
+	checkAdminConnexion();
 	$local_data = [getCoachProfile(),getOrderedPrefixPhone(), 'action_message' => 'coach', 'presentation_message' => 'Modification profil coach :' ];
 
 	display_view('admin/update_user_profile', $local_data);
@@ -263,6 +285,7 @@ function update_coach_profile () // show all the informations about a coach to u
 
 function update_coach_profile_proceed () // proceeds to the update of a coach profile
 {
+	checkAdminConnexion();
 	$oldInfo = getCoachProfile(); // get old information from database
 	$newInfo = getModificationCoachProfile(); // get new information from website (with update)
 	
@@ -306,6 +329,7 @@ function update_coach_profile_proceed () // proceeds to the update of a coach pr
 
 function delete_coach_profile_proceed () // proceeds the coach profile deletion
 {
+	checkAdminConnexion();
 	deleteCoachProfileProceed();
 
 	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=show_all_coachs_profile'); // redirect on the view to show coach profile with success message
@@ -313,20 +337,56 @@ function delete_coach_profile_proceed () // proceeds the coach profile deletion
 	setcookie('cookie_success_message', 'Le coach à bien été supprimé ! 💪', time() + 1, null, null, false, true); // cookie to set success message
 }
 
+//----------------------------- add coach -----------------------------
+
+function add_coach () // show coach form
+{
+	checkAdminConnexion();
+	$local_data = getPrefixPhone();
+
+	display_view('admin/add_coach', ['presentation_message' => "Formulaire d'ajout de coach :", $local_data]);
+
+}
+
+function add_coach_proceed () // show coach form
+{
+	checkAdminConnexion();
+	$local_data = getAddCoachInfo();
+
+	if (!empty(verifyExistingCoachMail($local_data['mail']))){
+		display_view('admin/add_coach', ['presentation_message' => "Formulaire d'ajout de coach :", 'error_message' => 'Cet email est déja utilisé par un autre coach.', $local_data]);
+		exit;
+	}
+	if (!empty(verifyExistingCoachPseudo($local_data['pseudo']))){
+		display_view('admin/add_coach', ['presentation_message' => "Formulaire d'ajout de coach :", 'error_message' => 'Ce pseudo est déja utilisé par un autre coach.', $local_data]);
+		exit;
+	}
+
+	addCoachProceed();
+
+	header('Location: https://srv-prj.iut-acy.local/RT/1projet17/mvc/public/index.php?controller=admin&action=show_all_coachs_profile'); // redirect on the view to show coach profile with success message
+
+	setcookie('cookie_success_message', 'Le coach à bien été ajouté ! 💪', time() + 1, null, null, false, true); // cookie to set success message
+
+}
+
 //----------------------------- show resources -----------------------------
 
 function show_resource_graph () // show graph
 {
+	checkAdminConnexion();
 	display_view('admin/show_resources', ['action_message'=>'graph', 'presentation_message' => 'Graphe des dépendances fonctionnelles']);
 }
 
 function show_resource_mcd () // show mcd
 {
+	checkAdminConnexion();
 	display_view('admin/show_resources', ['action_message'=>'mcd', 'presentation_message' => 'Modèle conceptuel de données']);
 }
 
 function show_resource_mld () // show mld
 {
+	checkAdminConnexion();
 	display_view('admin/show_resources', ['action_message'=>'mld', 'presentation_message' => 'Modèle logique de données']);
 }
 
@@ -334,6 +394,7 @@ function show_resource_mld () // show mld
 
 function show_changelog ()
 {
+	checkAdminConnexion();
 	display_view('admin/show_changelog', ['presentation_message' => 'Changelog DigitalSport :']);
 }
